+++ samples/app/src/main/java/com/example/app/tweetui/TweetListFragment.java
@@ -18,7 +18,6 @@
 package com.example.app.tweetui;
 
 import android.app.Activity;
-import android.content.Context;
 import android.os.Bundle;
 import android.support.v4.app.ListFragment;
 import android.view.LayoutInflater;
@@ -26,11 +25,13 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
+import com.twitter.sdk.android.core.Callback;
+import com.twitter.sdk.android.core.Result;
 import com.twitter.sdk.android.core.TwitterException;
 import com.twitter.sdk.android.core.models.Tweet;
-import com.twitter.sdk.android.tweetui.CompactTweetView;
-import com.twitter.sdk.android.tweetui.LoadCallback;
-import com.twitter.sdk.android.tweetui.TweetViewFetchAdapter;
+import com.twitter.sdk.android.tweetui.FixedTweetTimeline;
+import com.twitter.sdk.android.tweetui.TweetTimelineListAdapter;
+import com.twitter.sdk.android.tweetui.TweetUtils;
 
 import com.example.app.R;
 
@@ -56,17 +57,16 @@ public class TweetListFragment extends ListFragment {
         tweetIds.addAll(Arrays.asList(574000939800993792L, 503435417459249153L, 510908133917487104L,
                 473514864153870337L, 477788140900347904L, 20L, 484816434313195520L,
                 466041861774114819L, 448250020773380096L));
-        final TweetViewFetchAdapter adapter = new TweetViewFetchAdapter<CompactTweetView>(
-                getActivity()) {
+
+        TweetUtils.loadTweets(tweetIds, new Callback<List<Tweet>>() {
             @Override
-            public CompactTweetView getTweetView(Context context, Tweet tweet) {
-                return new CompactTweetView(context, tweet, R.style.tw__TweetLightStyle);
+            public void success(Result<List<Tweet>> result) {
+                final FixedTweetTimeline fixedTimeline = new FixedTweetTimeline.Builder()
+                        .setTweets(result.data).build();
+                final TweetTimelineListAdapter adapter = new TweetTimelineListAdapter(getActivity(),
+                        fixedTimeline);
+                setListAdapter(adapter);
             }
-        };
-        setListAdapter(adapter);
-        adapter.setTweetIds(tweetIds, new LoadCallback<List<Tweet>>() {
-            @Override
-            public void success(List<Tweet> tweets) {}
 
             @Override
             public void failure(TwitterException exception) {