+++ tweet-ui/src/main/java/com/twitter/sdk/android/tweetui/internal/TimelineDelegate.java
@@ -222,7 +222,7 @@ public class TimelineDelegate<T extends Identifiable> {
     }
 
     /**
-     * Handles receiving next timeline items. Prepends received items to listItems, update the
+     * Handles receiving next timeline items. Prepends received items to listItems, updates the
      * scrollStateHolder nextCursor, and calls notifyDataSetChanged.
      */
     class NextCallback extends DefaultCallback {